Split story files
